refactor: auth

+++ pkg/authentication/authenticated_user.go
@@ -1,7 +1,7 @@
 package authentication
 
 type AuthenticatedUser struct {
-	ID    uint64 `json:"id"`
+	ID    uint   `json:"id"`
 	Name  string `json:"name"`
 	Email string `json:"email"`
 }
